Backend/requirements.txt
django>=5.2.4
djangorestframework>=3.16.0
djangorestframework-simplejwt>=5.5.1
django-cors-headers>=4.7.0
django-filter>=23.2
mysqlclient>=2.2.7
pillow>=11.3.0
stripe>=12.3.0
python-dotenv>=1.0.0
